@@ -2,6 +2,7 @@
   "name": "web-client",
   "version": "0.0.1",
   "private": true,
+  "type": "module",
   "scripts": {
     "test": "echo ''",
     "dev": "next dev --turbopack",
@@ -9,8 +10,8 @@
     "export": "echo \"no action in script. Please run pnpm build (includes next export)\"",
     "start": "next start",
     "analyze": "ANALYZE=true next build",
-    "eslint-check": "eslint 'src/**/*.{ts,js,tsx,jsx}'",
-    "eslint-fix": "eslint --fix 'src/**/*.{ts,js,tsx,jsx}'",
+    "eslint-check": "eslint",
+    "eslint-fix": "eslint --fix",
     "lint": " pnpm eslint-check",
     "lint-fix": "pnpm eslint-fix",
     "clean": "rimraf build",
@@ -32,13 +33,13 @@
   "devDependencies": {
     "@eslint/compat": "^1.3.1",
     "@eslint/eslintrc": "^3.3.1",
-    "@eslint/js": "^9.29.0",
+    "@eslint/js": "^9.30.0",
     "@next/bundle-analyzer": "^15.3.4",
-    "@types/node": "^24.0.4",
+    "@types/node": "^24.0.7",
     "@types/react": "^19.1.8",
     "@types/react-dom": "^19.1.6",
     "autoprefixer": "^10.4.21",
-    "eslint": "^9.29.0",
+    "eslint": "^9.30.0",
     "eslint-config-import": "^0.13.0",
     "eslint-config-next": "^15.3.4",
     "eslint-import-resolver-typescript": "^4.4.4",
