@@ -1,9 +1,9 @@
 // eslint-disable-next-line @typescript-eslint/no-require-imports
-const withBundleAnalyzer = require('@next/bundle-analyzer')({
-  enabled: process.env.ANALYZE === 'true',
-});
+import withBundleAnalyzer from '@next/bundle-analyzer';
 
-module.exports = withBundleAnalyzer(
+export default withBundleAnalyzer({
+  enabled: process.env.ANALYZE === 'true',
+})(
   /** @type {import('next').NextConfig} */
   {
     output: 'export',
