@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   // add your custom config here
   // https://stylelint.io/user-guide/configuration
 }
